+++ package.json
@@ -25,7 +25,7 @@
     "@radix-ui/react-popover": "^1.1.7",
     "@radix-ui/react-separator": "^1.1.2",
     "@radix-ui/react-slot": "^1.1.2",
-    "@radix-ui/react-tabs": "^1.1.3",
+    "@radix-ui/react-tabs": "^1.1.4",
     "@radix-ui/react-tooltip": "^1.1.8",
     "@vercel/blob": "^0.27.1",
     "@vercel/speed-insights": "^1.2.0",
